+++ example/swagger/customerDefinition.ts
@@ -1,0 +1,6 @@
+export interface JsonResult<T> {
+  code: number
+  data: T
+  message: string
+  success: boolean
+}