+++ src/templates/genericTypeDefinitionTemplate.ts
@@ -26,4 +26,3 @@ export class PagedResultDto<T> implements IPagedResult<T> {
 }
 
 
-